refactor(common/sst): refactor drra component and controller classes, update includes and dependencies
