+++ lib/common/io_buffer/sst/io_buffer.cpp
@@ -3,6 +3,7 @@
 #include "custom_backing.h"
 
 #include "ioEvents.h"
+#include <sst/core/link.h>
 
 IOBuffer::IOBuffer(SST::ComponentId_t id, SST::Params &params) : Component(id) {
   // Output
