+++ lib/resources/iosram_top/sst/iosram_top.h
@@ -1,7 +1,7 @@
 #ifndef _IOSRAM_TOP_H
 #define _IOSRAM_TOP_H
 
-#include "drra.h"
+#include "drra_resource.h"
 
 #include "sst/elements/memHierarchy/membackend/backing.h"
